refactor: split http-server.ts into focused middleware and route modules
